Restart interrupted writes to stdout.
If some long text is output right after #help (such as another #help
command), SIGCHILD may cause the write() to return early.  Thus, we need
to restart the write (or at the very least not consider it a fatal error).

Seems to happen only in some circumstances, for example, never when the
output is redirected (thus it might be tricky to write a testsuite case).

Closes #3