Drop Mac OS builds from Travis.
This is especially bad as they're currently failing, and it'd be good to
debug that.  But, I don't have any Mac porterbox, and Travis runs cost
more than 25× as much as any other platform.  It's bad for the users,
but I'm not going to pay both in my efforts and in money to support a
platform whose vendor is aggressively hostile to free software developers.