Fix broken unaligned access in kbtree.
Per the C standard, the compiler may assume that such references never alias,
and optimize in a way that breaks kbtree.

It would be possible to use a compiler-specific way to mark these accesses
as unaligned, but for now, let's just waste 4 (64-bit) or 12 (128-bit) bytes
per node and use word alignment.

This isn't needed for keys smaller than a pointer, but these aren't used in
kbtin.