Add a configure check for getentropy().
While technically it is available on MacOS, there's some terrible fuckery
wrt XCode versions not matching libraries on the rest of the system, various
versions not working with each others, etc.  Reportedly it has been fixed
but at least whatever is installed on CirrusCI doesn't work.

According to a 2017 commit in libressl getentropy() is not functional on
at least iOS, and I'm not inclined to find out more.

Thus, let's not bother, it's not like KBtin uses randomness for anything
important.  That 1980's-style getpid()^time() trick is no regression.